refactor: no auth flow - migrate api service - use react query
