+++ src/components/auth/NoAuth/NoAuthFlow.tsx
@@ -1,10 +1,10 @@
-import { useEffect, useState } from 'react';
-import { GenerateConnectionRequest } from '@generated/api/src';
+import { useCallback } from 'react';
+import { GenerateConnectionOperationRequest } from '@generated/api/src';
 
-import { useApiKey } from 'context/ApiKeyContextProvider';
 import { useProject } from 'context/ProjectContextProvider';
-import { api, Connection } from 'services/api';
-import { handleServerError } from 'src/utils/handleServerError';
+import { Connection } from 'services/api';
+
+import { useCreateConnectionMutation } from '../useCreateConnectionMutation';
 
 import { NoAuthContent } from './NoAuthContent';
 
@@ -17,7 +17,6 @@ type NoAuthFlowProps = {
   providerName?: string;
   children: JSX.Element,
   selectedConnection: Connection | null;
-  setSelectedConnection: (connection: Connection | null) => void;
 };
 
 /**
@@ -27,36 +26,24 @@ type NoAuthFlowProps = {
  */
 export function NoAuthFlow({
   provider, consumerRef, consumerName, groupRef, groupName,
-  children, selectedConnection, setSelectedConnection, providerName,
+  children, selectedConnection, providerName,
 }: NoAuthFlowProps) {
-  const project = useProject();
-  const apiKey = useApiKey();
-  const [nextStep, setNextStep] = useState<boolean>(false);
+  const { projectIdOrName } = useProject();
+  const createConnectionMutation = useCreateConnectionMutation();
 
-  useEffect(() => {
-    if (provider && api && nextStep) {
-      const req: GenerateConnectionRequest = {
+  const onNext = useCallback(() => {
+    const req: GenerateConnectionOperationRequest = {
+      projectIdOrName,
+      generateConnectionParams: {
         groupName,
         groupRef,
         consumerName,
         consumerRef,
         provider,
-      };
-
-      api().connectionApi.generateConnection({ projectIdOrName: project.projectId, generateConnectionParams: req }, {
-        headers: { 'X-Api-Key': apiKey ?? '', 'Content-Type': 'application/json' },
-      }).then((conn) => {
-        setSelectedConnection(conn);
-      }).catch((err) => {
-        console.error('Error loading provider info.');
-        handleServerError(err);
-      });
-    }
-  }, [apiKey, provider, nextStep, consumerName, consumerRef, groupName, groupRef, setSelectedConnection, project]);
-
-  const onNext = () => {
-    setNextStep(true);
-  };
+      },
+    };
+    createConnectionMutation.mutate(req);
+  }, [projectIdOrName, groupName, groupRef, consumerName, consumerRef, provider, createConnectionMutation]);
 
   if (selectedConnection === null) {
     return <NoAuthContent handleSubmit={onNext} error={null} providerName={providerName} />;
